+++ EleFits/tests/src/BintableColumns_test.cpp
@@ -17,9 +17,9 @@
  *
  */
 
-#include "EleFitsData/TestColumn.h"
 #include "EleFits/BintableColumns.h"
 #include "EleFits/FitsFileFixture.h"
+#include "EleFitsData/TestColumn.h"
 
 #include <boost/test/unit_test.hpp>
 
@@ -78,7 +78,7 @@ BOOST_FIXTURE_TEST_CASE(append_rows_test, Test::TemporaryMefFile) {
   BOOST_TEST(columns.readRowCount() == initSize * 2);
 }
 
-template<typename T>
+template <typename T>
 void checkTupleWriteRead(const BintableColumns& du) {
   const long rowCount = 10000;
   const long repeatCount = 3;
@@ -99,27 +99,26 @@ void checkTupleWriteRead(const BintableColumns& du) {
   BOOST_TEST(res1.vector() == scalar.vector());
 }
 
-template<>
+template <>
 void checkTupleWriteRead<std::string>(const BintableColumns& du) {
   // FIXME (cannot use RandomVectorColumn)
   (void)du;
 }
 
-template<>
+template <>
 void checkTupleWriteRead<std::uint64_t>(const BintableColumns& du) {
   // FIXME CFitsIO bug?
   (void)du;
 }
 
-template<typename T>
-void checkVectorWriteRead(const BintableColumns& du) {
+template <typename T>
+void checkArrayWriteRead(const BintableColumns& du) {
   const long rowCount = 10000;
   const long repeatCount = 3;
-  std::vector<ColumnInfo<T>> infos { { "VECTOR", "m", repeatCount }, { "SCALAR", "s", 1 } }; // Inverted for robustness test
-  std::vector<VecColumn<T>> seq {
-      { infos[1], Test::generateRandomVector<T>(rowCount) },
-      { infos[0], Test::generateRandomVector<T>(repeatCount * rowCount) }
-  };
+  std::array<ColumnInfo<T>, 2> infos { ColumnInfo<T> { "VECTOR", "m", repeatCount },
+                                       ColumnInfo<T> { "SCALAR", "s", 1 } }; // Inverted for robustness test
+  std::array<VecColumn<T>, 2> seq { VecColumn<T> { infos[1], Test::generateRandomVector<T>(rowCount) },
+                                    VecColumn<T> { infos[0], Test::generateRandomVector<T>(repeatCount * rowCount) } };
   du.initSeq(0, infos);
   du.writeSeq(seq);
   const auto res = du.readSeq(Indexed<T>(0), Indexed<T>(1)); // TODO readSeq<T>({ 0, 1 }) and idem with strings?
@@ -131,13 +130,44 @@ void checkVectorWriteRead(const BintableColumns& du) {
   BOOST_TEST(res1.vector() == seq[0].vector());
 }
 
-template<>
+template <>
+void checkArrayWriteRead<std::string>(const BintableColumns& du) {
+  // FIXME (cannot use RandomVectorColumn)
+  (void)du;
+}
+
+template <>
+void checkArrayWriteRead<std::uint64_t>(const BintableColumns& du) {
+  // FIXME CFitsIO bug?
+  (void)du;
+}
+
+template <typename T>
+void checkVectorWriteRead(const BintableColumns& du) {
+  const long rowCount = 10000;
+  const long repeatCount = 3;
+  std::vector<ColumnInfo<T>> infos { { "VECTOR", "m", repeatCount },
+                                     { "SCALAR", "s", 1 } }; // Inverted for robustness test
+  std::vector<VecColumn<T>> seq { { infos[1], Test::generateRandomVector<T>(rowCount) },
+                                  { infos[0], Test::generateRandomVector<T>(repeatCount * rowCount) } };
+  du.initSeq(0, infos);
+  du.writeSeq(seq);
+  const auto res = du.readSeq(Indexed<T>(0), Indexed<T>(1)); // TODO readSeq<T>({ 0, 1 }) and idem with strings?
+  const auto& res0 = std::get<0>(res);
+  const auto& res1 = std::get<1>(res);
+  BOOST_TEST((res0.info() == seq[1].info()));
+  BOOST_TEST((res1.info() == seq[0].info()));
+  BOOST_TEST(res0.vector() == seq[1].vector());
+  BOOST_TEST(res1.vector() == seq[0].vector());
+}
+
+template <>
 void checkVectorWriteRead<std::string>(const BintableColumns& du) {
   // FIXME (cannot use RandomVectorColumn)
   (void)du;
 }
 
-template<>
+template <>
 void checkVectorWriteRead<std::uint64_t>(const BintableColumns& du) {
   // FIXME CFitsIO bug?
   (void)du;
@@ -145,12 +175,16 @@ void checkVectorWriteRead<std::uint64_t>(const BintableColumns& du) {
 
 #define SEQ_WRITE_READ_TEST(type, name) \
   BOOST_FIXTURE_TEST_CASE(name##_tuple_write_read_test, Test::TemporaryMefFile) { \
-    const auto& fromTuple = this->initBintableExt("TUPLE"); \
-    checkTupleWriteRead<type>(fromTuple.columns()); \
+    const auto& ext = this->initBintableExt("TUPLE"); \
+    checkTupleWriteRead<type>(ext.columns()); \
+  } \
+  BOOST_FIXTURE_TEST_CASE(name##_array_write_read_test, Test::TemporaryMefFile) { \
+    const auto& ext = this->initBintableExt("ARRAY"); \
+    checkArrayWriteRead<type>(ext.columns()); \
   } \
   BOOST_FIXTURE_TEST_CASE(name##_vector_write_read_test, Test::TemporaryMefFile) { \
-    const auto& fromVector = this->initBintableExt("VECTOR"); \
-    checkVectorWriteRead<type>(fromVector.columns()); \
+    const auto& ext = this->initBintableExt("VECTOR"); \
+    checkVectorWriteRead<type>(ext.columns()); \
   }
 
 ELEFITS_FOREACH_COLUMN_TYPE(SEQ_WRITE_READ_TEST)