+++ CHANGES.md
@@ -2,14 +2,13 @@
 
 ## 4.0.1
 
-### Bug fix
+### Bug fixes
 
-* `BintableColumns::initSeq()` was ill-formed
+* `BintableColumns::initSeq()` was ill-formed for non-tuple sequences
 
 ### New features
 
 * Empty binary table HDUs can be created
-* Homogeneous sequences of columns can be read and written (not only tuples)
 
 ## 4.0